requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
cffi==1.14.2
click==7.1.2
cycler==0.10.0
czifile==2019.7.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flake8==3.8.3
Flask==1.1.2
Flask-Sijax==0.4.1
future==0.18.2
imageio==2.9.0
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.1.3
numpy==1.19.0
opencv-python==4.3.0.36
openslide-python==1.1.1
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
read-roi==1.6.0
rope==0.17.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
Send2Trash==1.5.0
Sijax==0.3.2
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
